Prepare to fix bug in properties casting
The bug: properties outside of the schema must fail casting, and that works
for nested objects, but it doesn't work for root-level objects.

Unfortunately, one feature of cast/1 relies on this bug. My
strategy is to temporarily fork some of the code so that fixing the bug
doesn't break the feature.

The feature in question is the discriminator feature of Open API. I
forked the functions that are called by the code related to the feature.