Feat: add --check option in Mix tasks to compare generated spec w/ file
Add a --check option to run Mix tasks and compare the generated spec
with a previously generated file.

This is useful for scenarios where a CI check is desirable to catch
unwanted drifts from a validated OpenAPI spec: e.g. the OpenAPI spec is
committed and reviewed, and should not change without additional review.

Fixes #617

Signed-off-by: Davide Briani <davide@briani.dev>