fix: OTP28 compilation error and string test
Prevent storing regex in module attribute when calling `schema/2` macro,
instead perform function call to build the schema in `schema/0`.

This solution splits the build and validation steps of `build_schema/2`
as an optimization to prevent re-running the validation on every
`schema/0` call.

String test compares regex's sources instead of compiling new regex,
as this now fails because of mismatched `Reference`s.